added serialization with tests
