+++ python/semantic_kernel/connectors/ai/open_ai/services/azure_chat_completion.py
@@ -2,16 +2,18 @@
 
 
 from logging import Logger
-from typing import Optional
+from typing import Dict, Optional
 
-from pydantic import HttpUrl
+from pydantic import Field, HttpUrl
 
-from semantic_kernel.connectors.ai.open_ai.services.open_ai_chat_completion import (
+from semantic_kernel.connectors.ai.open_ai.services.open_ai_chat_completion_base import (
     OpenAIChatCompletionBase,
 )
 
 
 class AzureChatCompletion(OpenAIChatCompletionBase):
+    model_id: str = Field(..., alias="deployment_name")
+
     def __init__(
         self,
         deployment_name: str,
@@ -20,6 +22,7 @@ class AzureChatCompletion(OpenAIChatCompletionBase):
         api_version: Optional[str] = "2023-03-15-preview",
         ad_auth: bool = False,
         log: Optional[Logger] = None,
+        logger: Optional[Logger] = None,
     ) -> None:
         """
         Initialize an AzureChatCompletion service.
@@ -41,6 +44,7 @@ class AzureChatCompletion(OpenAIChatCompletionBase):
             ad_auth: Whether to use Azure Active Directory authentication. (Optional)
                 The default value is False.
             log: The logger instance to use. (Optional)
+            logger: deprecated, use 'log' instead.
         """
         kwargs = {
             "model_id": deployment_name,
@@ -49,6 +53,47 @@ class AzureChatCompletion(OpenAIChatCompletionBase):
             "api_version": api_version,
             "api_type": "azure_ad" if ad_auth else "azure",
         }
+        if logger:
+            logger.warning("The 'logger' argument is deprecated, use 'log' instead.")
+            kwargs["log"] = logger
         if log:
             kwargs["log"] = log
         super().__init__(**kwargs)
+
+    @classmethod
+    def from_dict(cls, settings: Dict[str, str]) -> "AzureChatCompletion":
+        """
+        Initialize an AzureChatCompletion service from a dictionary of settings.
+
+        Arguments:
+            settings: A dictionary of settings for the service.
+        """
+        if "api_type" in settings:
+            settings["ad_auth"] = settings["api_type"] == "azure_ad"
+            del settings["api_type"]
+
+        return AzureChatCompletion(
+            deployment_name=settings["deployment_name"],
+            endpoint=settings["endpoint"],
+            api_key=settings["api_key"],
+            api_version=settings.get("api_version"),
+            ad_auth=settings.get("ad_auth", False),
+            log=settings.get("log"),
+            # TODO: figure out if we need to be able to reinitialize the token counters.
+        )
+
+    def to_dict(self) -> Dict[str, str]:
+        """
+        Create a dict of the service settings.
+        """
+        # TODO: figure out if we need to be able to reinitialize the token counters.
+        return self.dict(
+            exclude={
+                "prompt_tokens",
+                "completion_tokens",
+                "total_tokens",
+                "api_type",
+                "org_id",
+            },
+            by_alias=True,
+        )
